+++ index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>GenkiV3</title>
-    <link rel="stylesheet" href="styles.css"/>
+    <link rel="stylesheet" href="./styles/layout.css"/>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Kumbh+Sans&display=swap" rel="stylesheet">
@@ -30,14 +30,6 @@
                     <a href="https://ebookcentral.proquest.com/lib/uql/detail.action?docID=6417968" target="_blank" class="navbar__links">Textbook</a>
                 </li>
             </ul>
-            <div class="dropdown">
-                <button onclick="myFunction()" class="dropbtn">Dropdown</button>
-                <div id="myDropdown" class="dropdown-content">
-                  <a href="#">Link 1</a>
-                  <a href="#">Link 2</a>
-                  <a href="#">Link 3</a>
-                </div>
-              </div> 
         </div>
     </nav>
 
@@ -91,6 +83,7 @@
     
     
 
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="./scripts/index.js"></script>
+    <script type="module" src="./scripts/settings.js"></script>
 </body>
 </html>
