+++ scripts/index.js
@@ -1,4 +1,4 @@
-import { KANJIONE, KANJITWO, GRAMMARTWO } from "./data.js";
+import { KANJIONE, KANJITWO, GRAMMARTWO } from "../data/data.js";
 
 const kanjiLength = KANJITWO.length;
 const ALLKANJI = KANJIONE.concat(KANJITWO)
@@ -6,26 +6,6 @@ const ALLKANJI = KANJIONE.concat(KANJITWO)
 function generateRandom(max) {
     return Math.floor(Math.random() * max) + 1;
 }
-
-/* When the user clicks on the button,
-toggle between hiding and showing the dropdown content */
-function myFunction() {
-    document.getElementById("myDropdown").classList.toggle("show");
-}
-  
-  // Close the dropdown menu if the user clicks outside of it
-  window.onclick = function(event) {
-    if (!event.target.matches('.dropbtn')) {
-      var dropdowns = document.getElementsByClassName("dropdown-content");
-      var i;
-      for (i = 0; i < dropdowns.length; i++) {
-        var openDropdown = dropdowns[i];
-        if (openDropdown.classList.contains('show')) {
-          openDropdown.classList.remove('show');
-        }
-      }
-    }
-  } 
 
 var darkMode = false
 var currArray = ALLKANJI
