--- styles.css
@@ -1,352 +1,0 @@
-* {
-    /*Box sizing allows padding and border properties
-    In the width and height of a div
-    By default, the actual width and height is calculated as
-    width + padding + border, thus using border-box would take
-    padding and border out of the actual width and height*/
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    font-family: 'Kumbh Sans', sans-serif;
-    /* turns out setting background colour in root ruins a lot of things...*/
-
-}
-
-/* How do I center this div */
-
-/*Navbar configuration*/
-.navbar {
-  background: #fff;
-  height: 80px;
-  /*flex brings elements onto a "row"
-  essentially positioning things like a box array*/
-  display: flex;
-  /*changes the alignment in a flex box
-  you can see the effect in https://developer.mozilla.org/en-US/docs/Web/CSS/justify-content*/
-  justify-content: center;
-  align-items: center;
-  /*font size relative to root font size, which acts as a "multipler"
-  by default the font size is 16px*/
-  font-size: 1.2rem;
-  position: sticky;
-  top: 0;
-  /*z-index is the "layer" of the element
-  999 or some big number ensures that the navbar is always above other
-  elements when scrolling through the site.*/
-  z-index: 999;
-}
-
-.navbar__container {
-  display: flex;
-  justify-content: space-between;
-  height: 80px;
-  z-index: 1;
-  width: 100%;
-  max-width: 1300px;
-  /*margin in px in order of top, right, bottom, left*/
-  margin: 0 auto;
-  padding: 0 50px;
-}
-
-/*# is used for html elements with id*/
-#navbar__logo {
-  background-color: #131313;
-  background-size: 100%;
-  /*These are used to make sure site works on different browsers*/
-  -webkit-background-clip: text;
-  -moz-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  -moz-text-fill-color: transparent;
-  /*Used for font of Logo*/
-  display: flex;
-  /*see below; center and stretch seems the most useful
-  https://developer.mozilla.org/en-US/docs/Web/CSS/align-items*/
-  align-items: center;
-  /*Determine what type of cursor is used when hovering over*/
-  cursor: pointer;
-  text-decoration: none;
-  font-size: 2rem;
-}
-
-.fa-gem {
-  margin-right: 0.5rem;
-}
-
-.navbar__menu {
-  display: flex;
-  align-items: center;
-  list-style: none;
-  text-align: center;
-}
-
-.navbar__item {
-  height: 80px;
-}
-
-.navbar__links {
-  color: black;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-decoration: none;
-  padding: 0 1rem;
-  height: 100%
-}
-
-.navbar__btn {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0 1rem;
-  width: 100%;
-}
-
-.navbar__links:hover {
-  color: #35721f;
-  transition: all 0.3s ease;
-}
-
-
- /* Dropdown Button */
- .dropbtn {
-  background-color: #3498DB;
-  color: white;
-  padding: 16px;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-}
-
-/* Dropdown button on hover & focus */
-.dropbtn:hover, .dropbtn:focus {
-  background-color: #2980B9;
-}
-
-/* The container <div> - needed to position the dropdown content */
-.dropdown {
-  position: relative;
-  display: inline-block;
-}
-
-/* Dropdown Content (Hidden by Default) */
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #f1f1f1;
-  min-width: 160px;
-  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-  z-index: 1;
-}
-
-/* Links inside the dropdown */
-.dropdown-content a {
-  color: black;
-  padding: 12px 16px;
-  text-decoration: none;
-  display: block;
-}
-
-/* Change color of dropdown links on hover */
-.dropdown-content a:hover {background-color: #ddd;}
-
-/* Show the dropdown menu (use JS to add this class to the .dropdown-content container when the user clicks on the dropdown button) */
-.show {display:block;} 
-
-
-/* END OF NAVBAR */
-
-
-.body {
-  display: flex;
-  justify-content: center;
-  background-color: #A8D897;
-}
-
-.kanji__container {
-    display: flex;
-    justify-content: center;
-    width: 60%;
-    z-index: 1;
-    padding-top: 2rem;
-}
-
-#kanji__element {
-  font-size: 50px;
-  font-weight: 500;
-  padding-left: 2rem;
-  padding-right: 2rem;
-  border-radius: 30px;
-  color: black;
-  background-color: transparent;
-  text-decoration: none;
-  text-align: center;
-}
-
-#kanji__element:hover {
-  background-color: #fff;
-  transition: all 0.3s ease;
-}
-
-#kanji__placeholder {
-  
-}
-
-/* END OF KANJI CONTAINER*/
-
-
-/* Generate button 
-
-Ripple effect is thanks to
-https://github.com/Basir-PD/100-Projects-HTML-CSS-JavaScript/blob/master/13-%20Button%20Ripple%20Effect/script.js
-*/
-
-.button {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #A8D897;
-}
-
-.button #generate {
-    font-size: 1rem;
-    background-image: linear-gradient(to top, #8d8d8d 0%, #d3d3d3 100%);
-    padding: 14px 32px;
-    border: none;
-    border-radius: 4px;
-    color: #141414;
-    margin-top: 2rem;
-    cursor: pointer;
-    position: relative;
-    transition: all 0.35s;
-    outline: none;
-}
-
-.btn-ripple {
-  position: relative;
-  background-image: linear-gradient(15deg, #13547a 0%, #80d0c7 100%);
-  color: #f3f3f3;
-  border-radius: 4rem;
-  overflow: hidden;
-  transition: box-shadow 0.3s ease-in-out;
-  cursor: pointer;
-}
-
-.btn-ripple:hover {
-  box-shadow: 0px 0px 0px 4px rgba(0, 0, 0, 0.1);
-}
-
-.ripple {
-  position: absolute;
-  background-color: rgba(0, 0, 0, 0.5);
-  transform: translate(-50%, -50%);
-  animation: rippleEffect 0.3s ease-out;
-  border-radius: 50%;
-}
-
-@keyframes rippleEffect {
-  from {
-    width: 0px;
-    height: 0px;
-  }
-  to {
-    width: 300px;
-    height: 300px;
-    opacity: 0;
-  }
-}
-
-.description {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #A8D897;
-}
-
-#desc {
-    text-align: center;
-}
-
-
-/* Create Hover */
-.chapter {
-  background-color: #A8D897;
-}
-
-.chapterselector h1 {
-  padding-top: 1rem;
-  text-align: center;
-}
-
-.chapterselector__menu {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  list-style: none;
-  text-align: center;
-  padding-top: 2rem;
-  margin: 0 30%;
-}
-
-.chapterselector__mobile {
-  display: none;
-}
-
-/* Problem with this is that the button scales with text size*/
-.chapterselector__item {
-  height: 60px;
-  width: 80px;
-  background-color: #d3d3d3;
-  padding: 15px 0;
-  border-radius: 20px;
-}
-
-.chapterdisplay {
-  display: flex;
-  justify-content: center;
-  margin-top: 30px;
-  margin-left: 20%;
-  width: 60%;
-  height: 300px;
-}
-
-/** All the flexbox stuff should be adjusted here*/
-/* Except I cannot get the box to expand fully here for some reason*/
-/* Why do you do what you do*/
-#grammar__container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  width: 100%;
-}
-
-#grammar__content {
-  background-color: #2196F3;
-  text-align: center;
-  font-size: 25px;
-  padding-top: 10px;
-  height: 20%;
-  width: 50%;
-  border-radius: 30px;
-}
-
-
-.footer {
-  padding-left: 2rem;
-  padding-top: 1rem;
-  padding-bottom: 1rem;
-}
-
-@media screen and (max-width: 960px) {
-  /* 2 bars collapse into 1, with a little bit of spacing*/
-  #grammar__content {
-    width: 100%;
-    font-size: 20px;
-  }
-
-  /* this will be displayed as a drop box instead */
-  .chapterselector__menu {
-    display: none;
-  }
-}
