+++ styles/Grammar.css
@@ -1,0 +1,87 @@
+
+
+
+
+/* Create Hover */
+.chapter {
+    background-color: #A8D897;
+  }
+  
+  .chapterselector h1 {
+    padding-top: 1rem;
+    text-align: center;
+  }
+  
+  .chapterselector__menu {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    list-style: none;
+    text-align: center;
+    padding-top: 2rem;
+    margin: 0 30%;
+  }
+  
+  .chapterselector__mobile {
+    display: none;
+  }
+  
+  /* Problem with this is that the button scales with text size*/
+  .chapterselector__item {
+    height: 60px;
+    width: 80px;
+    background-color: #d3d3d3;
+    padding: 15px 0;
+    border-radius: 20px;
+  }
+  
+  .chapterdisplay {
+    display: flex;
+    justify-content: center;
+    margin-top: 30px;
+    margin-left: 20%;
+    width: 60%;
+    height: 300px;
+  }
+  
+  /** All the flexbox stuff should be adjusted here*/
+  /* Except I cannot get the box to expand fully here for some reason*/
+  /* Why do you do what you do*/
+  #grammar__container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    width: 100%;
+  }
+  
+  #grammar__content {
+    background-color: #2196F3;
+    text-align: center;
+    font-size: 25px;
+    padding-top: 10px;
+    height: 20%;
+    width: 50%;
+    border-radius: 30px;
+  }
+  
+  
+  .footer {
+    padding-left: 2rem;
+    padding-top: 1rem;
+    padding-bottom: 1rem;
+  }
+  
+  @media screen and (max-width: 960px) {
+    /* 2 bars collapse into 1, with a little bit of spacing*/
+    #grammar__content {
+      width: 100%;
+      font-size: 20px;
+    }
+  
+    /* this will be displayed as a drop box instead */
+    .chapterselector__menu {
+      display: none;
+    }
+  }
