+++ styles/KanjiBody.css
@@ -1,0 +1,115 @@
+/* How do I center this div */
+
+
+
+
+.body {
+    display: flex;
+    justify-content: center;
+    background-color: #A8D897;
+  }
+  
+  .kanji__container {
+      display: flex;
+      justify-content: center;
+      width: 60%;
+      z-index: 1;
+      padding-top: 2rem;
+  }
+  
+  #kanji__element {
+    font-size: 50px;
+    font-weight: 500;
+    padding-left: 2rem;
+    padding-right: 2rem;
+    border-radius: 30px;
+    color: black;
+    background-color: transparent;
+    text-decoration: none;
+    text-align: center;
+  }
+  
+  #kanji__element:hover {
+    background-color: #fff;
+    transition: all 0.3s ease;
+  }
+  
+  #kanji__placeholder {
+    
+  }
+  
+  /* END OF KANJI CONTAINER*/
+  
+  
+  /* Generate button 
+  
+  Ripple effect is thanks to
+  https://github.com/Basir-PD/100-Projects-HTML-CSS-JavaScript/blob/master/13-%20Button%20Ripple%20Effect/script.js
+  */
+  
+  .button {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      background-color: #A8D897;
+  }
+  
+  .button #generate {
+      font-size: 1rem;
+      background-image: linear-gradient(to top, #8d8d8d 0%, #d3d3d3 100%);
+      padding: 14px 32px;
+      border: none;
+      border-radius: 4px;
+      color: #141414;
+      margin-top: 2rem;
+      cursor: pointer;
+      position: relative;
+      transition: all 0.35s;
+      outline: none;
+  }
+  
+  .btn-ripple {
+    position: relative;
+    background-image: linear-gradient(15deg, #13547a 0%, #80d0c7 100%);
+    color: #f3f3f3;
+    border-radius: 4rem;
+    overflow: hidden;
+    transition: box-shadow 0.3s ease-in-out;
+    cursor: pointer;
+  }
+  
+  .btn-ripple:hover {
+    box-shadow: 0px 0px 0px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .ripple {
+    position: absolute;
+    background-color: rgba(0, 0, 0, 0.5);
+    transform: translate(-50%, -50%);
+    animation: rippleEffect 0.3s ease-out;
+    border-radius: 50%;
+  }
+  
+  @keyframes rippleEffect {
+    from {
+      width: 0px;
+      height: 0px;
+    }
+    to {
+      width: 300px;
+      height: 300px;
+      opacity: 0;
+    }
+  }
+  
+  .description {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      background-color: #A8D897;
+  }
+  
+  #desc {
+      text-align: center;
+  }
+
